+++ applications/vessel-history/src/features/profile/components/Info.tsx
@@ -238,38 +238,40 @@ const Info: React.FC<InfoProps> = (props): React.ReactElement => {
                 valuesHistory={vessel.history.operator.byDate}
               ></InfoField>
 
-              <div className={styles.identifierField}>
-                <label>{t(`vessel.aisTransmission_plural`, 'AIS Transmissions')}</label>
-                <div>
-                  {vessel.firstTransmissionDate || vessel.lastTransmissionDate ? (
-                    <Fragment>
-                      {t('common.from', 'from')}{' '}
-                      {vessel.firstTransmissionDate ? (
-                        <I18nDate date={vessel.firstTransmissionDate} />
-                      ) : (
-                        DEFAULT_EMPTY_VALUE
-                      )}{' '}
-                      {t('common.to', 'to')}{' '}
-                      {vessel.lastTransmissionDate ? (
-                        <I18nDate date={vessel.lastTransmissionDate} />
-                      ) : (
-                        DEFAULT_EMPTY_VALUE
-                      )}
-                    </Fragment>
-                  ) : (
-                    DEFAULT_EMPTY_VALUE
-                  )}
+              {!isMergedVesselsView && (
+                <div className={styles.identifierField}>
+                  <label>{t(`vessel.aisTransmission_plural`, 'AIS Transmissions')}</label>
+                  <div>
+                    {vessel.firstTransmissionDate || vessel.lastTransmissionDate ? (
+                      <Fragment>
+                        {t('common.from', 'from')}{' '}
+                        {vessel.firstTransmissionDate ? (
+                          <I18nDate date={vessel.firstTransmissionDate} />
+                        ) : (
+                          DEFAULT_EMPTY_VALUE
+                        )}{' '}
+                        {t('common.to', 'to')}{' '}
+                        {vessel.lastTransmissionDate ? (
+                          <I18nDate date={vessel.lastTransmissionDate} />
+                        ) : (
+                          DEFAULT_EMPTY_VALUE
+                        )}
+                      </Fragment>
+                    ) : (
+                      DEFAULT_EMPTY_VALUE
+                    )}
+                  </div>
                 </div>
-              </div>
+              )}
               <InfoField
                 vesselName={vessel.shipname ?? DEFAULT_EMPTY_VALUE}
                 label={VesselFieldLabel.iuuStatus}
                 value={
                   vessel.iuuStatus !== undefined
                     ? t(
-                      `vessel.iuuStatusOptions.${vessel.iuuStatus}` as any,
-                      vessel.iuuStatus.toString()
-                    )
+                        `vessel.iuuStatusOptions.${vessel.iuuStatus}` as any,
+                        vessel.iuuStatus.toString()
+                      )
                     : DEFAULT_EMPTY_VALUE
                 }
                 valuesHistory={[]}