+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.70
+
+### Patch Changes
+
+- 48c18a7d: support custom style
+
+## 1.1.69
+
+### Patch Changes
+
+- Updated dependencies [3ce0364b]
+  - @globalfishingwatch/ui-components@12.4.1
+
+## 1.1.68
+
+### Patch Changes
+
+- Updated dependencies [8978b00d]
+  - @globalfishingwatch/timebar@2.3.0
+
+## 1.1.67
+
+### Patch Changes
+
+- Updated dependencies [0dd8b81d]
+  - @globalfishingwatch/i18n-labels@1.2.0
+  - @globalfishingwatch/timebar@2.2.0
+
+## 1.1.66
+
+### Patch Changes
+
+- Updated dependencies [b88273847]
+  - @globalfishingwatch/layer-composer@10.0.2
+
+## 1.1.65
+
+### Patch Changes
+
+- Updated dependencies [24642e15]
+  - @globalfishingwatch/ui-components@12.4.0
+
+## 1.1.64
+
+### Patch Changes
+
+- Updated dependencies [39c43baa]
+  - @globalfishingwatch/ui-components@12.3.1
+
 ## 1.1.63
 
 ### Patch Changes