+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.142
+
+### Patch Changes
+
+- Updated dependencies [8978b00d]
+  - @globalfishingwatch/timebar@2.3.0
+
+## 0.7.141
+
+### Patch Changes
+
+- Updated dependencies [0dd8b81d]
+  - @globalfishingwatch/timebar@2.2.0
+
+## 0.7.140
+
+### Patch Changes
+
+- Updated dependencies [b88273847]
+  - @globalfishingwatch/layer-composer@10.0.2
+
 ## 0.7.139
 
 ### Patch Changes